5488: Checks should use proxy when configured
In the AWS environment applications must use an egress proxy to reach
www.hub-dr.signin.service.gov.uk. The sensu metadata checks running on
the apps must also use a proxy. The easiest way to configure sensu to
use a proxy for the checks is to use the `http_proxy` environment
variable. `Net::HTTP` recognises this variable and uses it. However
one check uses `open-uri` which, although documented that it should,
does not pick up the http_proxy. We also found a case where Net::HTTP is
but does not pick up the http_proxy variable.

We have therefore changed the first case to use Net::HTTP and modified
the second case so that it creates and instance of the HTTP object. Both
now pick up the http_proxy variable.

Pair: @nickhaddock @edrowley