Add metrics for OCSP success
there was a problem with r509-ocsp-responder 0.3.3 which didn't work
with the ruby::openssl in modern ruby versions (>= 2.4).  There is a
fix [1] but, although the fix was made in 2017, it still hasn't been
released to rubygems, so we have to pull it in via git.

also, we require bundler 2.x now, ruby:2.5.3 comes with bundler 1.17.2

[1]: https://github.com/r509/r509-ocsp-responder/commit/1ffdfd530548d3ed9099ac0ff30825fd79c961ac