Accept metadata as a single directory
This merges the --metadata_cas and --cas parameters into one.  The new
--cas parameter takes a directory instead of a list of files.

This means that we are no longer as strict as we were: previously we
meticulously checked that each certificate chained back to an
appropriate CA.  But this isn't needed for the stated metrics of
metadata expiry, certificate expiry, or OCSP status checking.

The apps themselves check that certificates are correctly chaining to
the right CA, so we don't need to do this check here.

I also:

 - removed the metadata_expected_to_be_signed parameters from
   ValidUntilMetric and ExpiryDateMetric. ValidUntilMetric didn't use
   it, and we think that we sign metadata in all environments, so we
   don't need to have any logic here.
 - changed the `-h HOST` parameter to `-m METADATA_URL` to be clearer
   as to what it actually is