rename metrics from verify_federation_* to verify_metadata_*
"Federation" is not a good name. It probably originates from the use
of `/federation` in the metadata URL, but since metadata doesn't
contain all the certificates in the federation (it doesn't include RP
certs), this confuses people.

If we say `metadata` instead, it will be clearer what we mean.

After this change, we'll have to update our grafana dashboards but
this is easier to do now than later.