HUB-1013: Check metadata sources against deployed
This adds a new prometheus gauge which will generated metrics with
information on differences between the certificates defined in the
metadata repo and those published in the live metadata.

This is being done as part of the monitoring work for signing production
metadata in the pipeline.

These metrics can feed into an alert configured in Grafana.