HUB-1013: Use Faraday http cache
Octokit can be configured to use a Faraday http cache to cache
resources. This results in less calls to the GitHub API and so we won't
get anywhere near the API rate limit.

The middleware will use the ETag of the resources to determine if a file
has changed and if so it will pull it again.

This is documented in the Octokit docs here: https://github.com/octokit/octokit.rb#caching

This also fixes an issue where certs pulled from the published metadata
were being compared against the source certs without removing newlines.
This caused the metric to falsley report differences.