BAU: Handle OCSP errors
If the OCSP checking endpoint is down, as has been recently, the OCSP
client throws an error.

This error is unhandled and causes the metadata-exporter process to fail
to report any metrics at all.

This handles the exception and just logs the reason. This should allow
the process to continue and report other metrics successfully.